+++ test/connected.jl
@@ -19,7 +19,7 @@
     @test label_components(A, connectivity) == lbltarget2
     @test component_boxes(lbltarget) == Vector{Tuple}[[(1,2),(2,3)],[(1,1),(2,2)],[(1,3),(2,4)]]
     @test component_lengths(lbltarget) == [2,3,3]
-    @test component_indices(lbltarget) == Array{Int64}[[4,5],[1,2,3],[6,7,8]]
+    @test component_indices(lbltarget) == Array{Int}[[4,5],[1,2,3],[6,7,8]]
     @test component_subscripts(lbltarget) == Array{Tuple}[[(2,2),(1,3)],[(1,1),(2,1),(1,2)],[(2,3),(1,4),(2,4)]]
     @test @inferred(component_centroids(lbltarget)) == Tuple[(1.5,2.5),(4/3,4/3),(5/3,11/3)]
